fix: add cache eviction for tag search
